backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
supabase>=2.0.0
httpx>=0.25.0,<1.0.0
python-multipart>=0.0.12,<1.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
email-validator>=2.1.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
slowapi>=0.1.9,<1.0.0
reportlab>=4.1.0,<5.0.0
qrcode[pil]>=7.4.2,<8.0.0
Pillow>=10.3.0,<11.0.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
numpy>=1.24.0,<2.0.0
opencv-python>=4.7.0,<5.0.0
librosa>=0.10.0,<1.0.0
soundfile>=0.12.0,<1.0.0
torch>=2.0.0
torchvision>=0.15.0
